requirements.txt
Pillow==3.0.0
argparse==1.2.1
numpy==1.10.1
wsgiref==0.1.2
